requirements.txt
# Core FastAPI and server
fastapi==0.118.0
uvicorn[standard]==0.37.0
pydantic==2.11.9

# GitHub integration
PyGithub==2.8.1
requests==2.32.5

# OpenAI integration
openai==1.109.1

# HTTP client
httpx==0.28.1

# Environment management
python-dotenv==1.1.1

# Security and crypto
cryptography==46.0.1
PyJWT==2.10.1
